feat: move widgets to separated files
